refactor(container): mv MailBox to tool scope
