+++ src/containers/tool/MailBox/styles/mails_panel.js
