+++ src/containers/unit/Header/Header.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import MailBox from '@/containers/MailBox'
+import MailBox from '@/containers/tool/MailBox'
 import UserLister from '@/containers/user/UserLister'
 // import Cashier from '@/containers/tool/Cashier'
 // import UpgradePackages from '@/containers/tool/UpgradePackages'
